+++ R/randtest.dist.R
@@ -2,7 +2,7 @@
 #'
 #' @description Measures the distance between the observed statistic from a \code{"randtest"} object and some specific quantile of the simulated data.
 #'
-#' @param randtest an object of class \code{"randtest"}
+#' @param xtest an object of class \code{"randtest"}
 #' @param quantile a \code{numeric} value for the quantile edges to compare the observed data to on either sides (by default \code{quantile = c(0.025. 0.975)}).
 #' @param abs \code{logical}, whether to calculate the distance as an absolute value (\code{TRUE}) or not (\code{FALSE} - default).
 #' 
@@ -32,7 +32,9 @@
 #' @author Thomas Guillerme
 #' @export
 
-randtest.dist <- function(randtest, quantile = c(0.025, 0.975), abs = FALSE) {
+randtest.dist <- function(xtest, quantile = c(0.025, 0.975), abs = FALSE) {
+
+    randtest <- xtest
 
     ## Checking randtest
     check.class(randtest, "randtest")